requirements.txt
alabaster==0.7.12
attrs==20.3.0
Babel==2.9.0
branca==0.4.1
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
click-plugins==1.1.1
cligj==0.7.0
coloredlogs==14.0
docutils==0.16
earthengine-api==0.1.238
Fiona==1.8.17
folium==0.11.0
future==0.18.2
geopandas==0.8.1
google-api-core==1.22.4
google-api-python-client==1.12.3
google-auth==1.22.1
google-auth-httplib2==0.0.4
google-cloud-core==1.4.3
google-cloud-storage==1.31.2
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
httplib2==0.18.1
httplib2shim==0.0.3
humanfriendly==8.2
idna==2.10
imagesize==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
munch==2.5.0
numpy==1.19.4
packaging==20.4
pandas==1.1.4
Pillow==7.2.0
pkg-resources==0.0.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycrypto==2.6.1
Pygments==2.7.2
pyparsing==2.4.7
pyproj==3.0.0.post1
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
requests==2.24.0
rsa==4.6
Shapely==1.7.1
six==1.15.0
snowballstemmer==2.0.0
Sphinx==3.3.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
uritemplate==3.0.1
urllib3==1.25.10
